+++ pkg/config/admin_models.go
@@ -1,0 +1,5 @@
+package config
+
+type AdminConfig struct {
+	DefaultList []string `yaml:"defaultList"`
+}