+++ pkg/services/admin/admin_management_service_test.go
@@ -2,6 +2,7 @@ package admin
 
 import (
 	apiUtils "github.com/RoadTripMoustache/iris_api/pkg/apirouter/utils"
+	"github.com/RoadTripMoustache/iris_api/pkg/config"
 	dbmodels "github.com/RoadTripMoustache/iris_api/pkg/dbmodels/admin"
 	"github.com/RoadTripMoustache/iris_api/pkg/errors"
 	"github.com/RoadTripMoustache/iris_api/pkg/mocks/services"
@@ -9,6 +10,20 @@ import (
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
+
+var NoAdminConfig = config.Config{
+	Admin: config.AdminConfig{
+		DefaultList: []string{},
+	},
+}
+
+var WithAdminConfig = config.Config{
+	Admin: config.AdminConfig{
+		DefaultList: []string{
+			"toto@toto.com",
+		},
+	},
+}
 
 // ----- GetAdmin ----- //
 func Test_GetAdmin(t *testing.T) {
@@ -18,45 +33,76 @@ func Test_GetAdmin(t *testing.T) {
 		queryParams    map[string][]string
 		expectedResult *dbmodels.Admin
 		exprectedErr   *errors.EnhancedError
+		mockedConfig   config.Config
 	}{
 		{
-			caseDesc:       "Empty DB",
+			caseDesc:       "Empty DB and no default admin",
 			dbResult:       []map[string]interface{}{},
 			queryParams:    nil,
 			expectedResult: nil,
 			exprectedErr:   nil,
+			mockedConfig:   NoAdminConfig,
 		},
 		{
-			caseDesc:    "Not empty DB",
+			caseDesc:    "Not empty DB and no default admin",
 			queryParams: nil,
 			dbResult: []map[string]interface{}{
 				{
-					"user_id": "toto",
+					"user_email": "toto@toto.com",
 				},
 			},
 			expectedResult: &dbmodels.Admin{
-				UserID: "toto",
+				UserEmail: "toto@toto.com",
 			},
 			exprectedErr: nil,
+			mockedConfig: NoAdminConfig,
+		},
+		{
+			caseDesc:    "Empty DB and with default admin",
+			dbResult:    []map[string]interface{}{},
+			queryParams: nil,
+			expectedResult: &dbmodels.Admin{
+				UserEmail: "toto@toto.com",
+			},
+			exprectedErr: nil,
+			mockedConfig: WithAdminConfig,
+		},
+		{
+			caseDesc:    "Not empty DB and with default admin",
+			queryParams: nil,
+			dbResult: []map[string]interface{}{
+				{
+					"user_email": "toto@toto.com",
+				},
+			},
+			expectedResult: &dbmodels.Admin{
+				UserEmail: "toto@toto.com",
+			},
+			exprectedErr: nil,
+			mockedConfig: WithAdminConfig,
 		},
 	}
 
 	for _, testCase := range testCases {
 		t.Run(testCase.caseDesc, func(t *testing.T) {
 			// --- Given
-			userID := "toto_uuid"
+			userEmail := "toto@toto.com"
 			ctx := apiUtils.Context{
-				UserID:      userID,
+				UserEmail:   userEmail,
 				QueryParams: testCase.queryParams,
+			}
+
+			configGetConfigs = func() config.Config {
+				return testCase.mockedConfig
 			}
 
 			mNoSQLStorageInstance := new(services.MockNoSQLStorageInstance)
 			mNoSQLStorageInstance.
 				On("GetFirstDocument", dbmodels.AdminCollectionName,
 					[]utils.Filter{{
-						Param:    dbmodels.AdminUserIDLabel,
+						Param:    dbmodels.AdminUserEmailLabel,
 						Operator: "==",
-						Value:    "totoU",
+						Value:    "toto@toto.com",
 					}}).
 				Return(testCase.dbResult)
 
@@ -65,11 +111,13 @@ func Test_GetAdmin(t *testing.T) {
 			noSQLStorageGetInstance = mockNoSQLStorageService.GetInstance
 
 			// --- When
-			result, err := GetAdmin(ctx, "totoU")
+			result, err := GetAdmin(ctx, "toto@toto.com")
 
 			// --- Then
-			mNoSQLStorageInstance.AssertExpectations(t)
-			mockNoSQLStorageService.AssertExpectations(t)
+			if len(testCase.mockedConfig.Admin.DefaultList) == 0 {
+				mNoSQLStorageInstance.AssertExpectations(t)
+				mockNoSQLStorageService.AssertExpectations(t)
+			}
 			assert.Equal(t, testCase.exprectedErr, err)
 			assert.Equal(t, testCase.expectedResult, result)
 		})
@@ -84,60 +132,101 @@ func Test_GetAdmins(t *testing.T) {
 		queryParams    map[string][]string
 		expectedResult []*dbmodels.Admin
 		exprectedErr   *errors.EnhancedError
+		mockedConfig   config.Config
 	}{
 		{
-			caseDesc:       "Empty DB",
+			caseDesc:       "Empty DB and no default admin",
 			dbResult:       []map[string]interface{}{},
 			queryParams:    nil,
 			expectedResult: nil,
 			exprectedErr:   nil,
+			mockedConfig:   NoAdminConfig,
 		},
 		{
-			caseDesc:    "1 item",
+			caseDesc:    "1 item in DB",
 			queryParams: nil,
 			dbResult: []map[string]interface{}{
 				{
-					"user_id": "toto",
+					"user_email": "toto",
 				},
 			},
 			expectedResult: []*dbmodels.Admin{
 				{
-					UserID: "toto",
+					UserEmail: "toto",
 				},
 			},
 			exprectedErr: nil,
+			mockedConfig: NoAdminConfig,
 		},
 		{
-			caseDesc:    "2 items",
+			caseDesc:    "2 items in DB",
 			queryParams: nil,
 			dbResult: []map[string]interface{}{
 				{
-					"user_id": "toto",
+					"user_email": "toto",
 				},
 				{
-					"user_id": "titi",
+					"user_email": "titi",
 				},
 			},
 			expectedResult: []*dbmodels.Admin{
 				{
-					UserID: "toto",
+					UserEmail: "toto",
 				},
 				{
-					UserID: "titi",
+					UserEmail: "titi",
 				},
 			},
 			exprectedErr: nil,
+			mockedConfig: NoAdminConfig,
+		},
+		{
+			caseDesc:    "1 item in default admin",
+			queryParams: nil,
+			dbResult:    nil,
+			expectedResult: []*dbmodels.Admin{
+				{
+					UserEmail: "toto@toto.com",
+				},
+			},
+			exprectedErr: nil,
+			mockedConfig: WithAdminConfig,
+		},
+		{
+			caseDesc:    "2 items in DB and default admin",
+			queryParams: nil,
+			dbResult: []map[string]interface{}{
+				{
+					"user_email": "toto",
+				},
+				{
+					"user_email": "titi",
+				},
+			},
+			expectedResult: []*dbmodels.Admin{
+				{
+					UserEmail: "toto",
+				},
+				{
+					UserEmail: "titi",
+				},
+				{
+					UserEmail: "toto@toto.com",
+				},
+			},
+			exprectedErr: nil,
+			mockedConfig: WithAdminConfig,
 		},
 	}
 
 	for _, testCase := range testCases {
 		t.Run(testCase.caseDesc, func(t *testing.T) {
 			// --- Given
-			userID := "toto_uuid"
+			userEmail := "toto@toto.com"
 			pageNumber := 1
 			pageOffset := 0
 			ctx := apiUtils.Context{
-				UserID:      userID,
+				UserEmail:   userEmail,
 				QueryParams: testCase.queryParams,
 				Pagination: apiUtils.Pagination{
 					PageNumber: &pageNumber,
@@ -155,6 +244,10 @@ func Test_GetAdmins(t *testing.T) {
 			mockNoSQLStorageService.On("GetInstance").Return(mNoSQLStorageInstance)
 			noSQLStorageGetInstance = mockNoSQLStorageService.GetInstance
 
+			configGetConfigs = func() config.Config {
+				return testCase.mockedConfig
+			}
+
 			// --- When
 			result, err := GetAdmins(ctx)
 
@@ -171,10 +264,10 @@ func Test_GetAdmins(t *testing.T) {
 func Test_AddAdmin(t *testing.T) {
 	// --- Given
 	ctx := apiUtils.Context{
-		UserID: "toto",
+		UserEmail: "toto",
 	}
 	expectedResult := dbmodels.Admin{
-		UserID: "titi",
+		UserEmail: "titi",
 	}
 
 	mNoSQLStorageInstance := new(services.MockNoSQLStorageInstance)
@@ -186,7 +279,7 @@ func Test_AddAdmin(t *testing.T) {
 	mNoSQLStorageInstance.
 		On("GetFirstDocument", dbmodels.AdminCollectionName,
 			[]utils.Filter{{
-				Param:    dbmodels.AdminUserIDLabel,
+				Param:    dbmodels.AdminUserEmailLabel,
 				Operator: "==",
 				Value:    "titi",
 			}}).
@@ -213,7 +306,7 @@ func Test_DeleteAdmin(t *testing.T) {
 	mNoSQLStorageInstance := new(services.MockNoSQLStorageInstance)
 	var nilError error
 	mNoSQLStorageInstance.
-		On("Delete", dbmodels.AdminCollectionName, "titi", dbmodels.AdminUserIDLabel).
+		On("Delete", dbmodels.AdminCollectionName, "titi", dbmodels.AdminUserEmailLabel).
 		Return(nilError)
 
 	mockNoSQLStorageService := new(services.MockNoSQLStorage)